backend/requirements.txt
Flask==2.2.5
flask_cors==6.0.1
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
alembic==1.16.4
requests==2.28.1
sqlalchemy==2.0.42
psycopg2-binary==2.9.9
boto3==1.38.0
Werkzeug==2.3.8
PyJWT==2.8.0
